+++ Infrastructure/BaseClasses/BaseService.cs
@@ -1,5 +1,6 @@
 ﻿namespace Infrastructure.BaseClasses
 {
+    using System;
     using System.Collections.Generic;
 
     using Infrastructure.Messaging.Requests;
@@ -14,12 +15,24 @@
     {
         private readonly TRepository repository;
 
+        protected bool IsGetMethodAllowed { get; set; }
+        protected bool IsGetAllMethodAllowed { get; set; }
+        protected bool IsAddMethodAllowed { get; set; }
+        protected bool IsUpdateMethodAllowed { get; set; }
+        protected bool IsRemoveMethodAllowed { get; set; }
+
         /// <summary>
         /// Constructor
         /// </summary>
         protected BaseService(TRepository repository)
         {
             this.repository = repository;
+
+            IsGetMethodAllowed = true;
+            IsGetAllMethodAllowed = true;
+            IsAddMethodAllowed = true;
+            IsUpdateMethodAllowed = true;
+            IsRemoveMethodAllowed = true;
         }
 
         /// <summary>
@@ -27,6 +40,11 @@
         /// </summary>
         public Response<TModel> Get(GetRequest request)
         {
+            if (!IsGetMethodAllowed)
+            {
+                throw new InvalidOperationException("The service doesn't allow this method");
+            }
+
             TModel model = repository.GetById(request.Id);
 
             if (model == null)
@@ -42,6 +60,11 @@
         /// </summary>
         public Response<List<TModel>> GetAll(GetAllRequest request)
         {
+            if (!IsGetAllMethodAllowed)
+            {
+                throw new InvalidOperationException("The service doesn't allow this method");
+            }
+
             List<TModel> models = repository.GetAll(request.PageIndex, request.PageSize);
 
             return Response<List<TModel>>.CreateSuccess(models);
@@ -52,6 +75,11 @@
         /// </summary>
         public Response<TModel> Add(AddRequest<TModel> request)
         {
+            if (!IsAddMethodAllowed)
+            {
+                throw new InvalidOperationException("The service doesn't allow this method");
+            }
+
             TModel model = repository.Add(request.Model);
 
             if (model == null)
@@ -67,6 +95,11 @@
         /// </summary>
         public Response<TModel> Update(UpdateRequest<TModel> request)
         {
+            if (!IsUpdateMethodAllowed)
+            {
+                throw new InvalidOperationException("The service doesn't allow this method");
+            }
+
             TModel model = repository.Update(request.Model);
 
             if (model == null)
@@ -82,6 +115,11 @@
         /// </summary>
         public Response<Empty> Remove(RemoveRequest request)
         {
+            if (!IsRemoveMethodAllowed)
+            {
+                throw new InvalidOperationException("The service doesn't allow this method");
+            }
+
             repository.Remove(request.Id);
 
             return Response<Empty>.CreateSuccess();
